[modify] 맵 오브젝트 나누기 완료. 라이트맵 수정 완료
맵의 오브젝트들을 각각 나누어서 UV를 펼치는 작업 완료.
라이트맵 Bake 진행 시 빛이 이상하게 보이는 현상 수정.

+++ Project109/Assets/Scripts/Map/BattleMapManager.cs
@@ -21,7 +21,7 @@ public class BattleMapManager : MonoBehaviour
     public (int column, int row) centerCoord;
     public (int column, int row) rewardCoord;
     public (int column, int row) playerSpawnCoord;
-    public int tilePadding;
+    public float tilePadding;
 
     public Vector3 CheckTileMapLocationByRowAndColumn(int newColumn, int newRow)
     {
